packages/cli/pyproject.toml
[project]
name = "memex-cli"
version = "0.0.1"
description = "Memex CLI provides command-line tools for interacting with the Memex system."
authors = [
  {name="Jasper Ginn", email="jasperginn@gmail.com"}
]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "memex-common",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "platformdirs>=4.2.0",
    "python-box>=7.1.1",
    "aiofiles>=25.1.0",
]

[project.optional-dependencies]
dashboard = ["memex_dashboard"]
server = ["memex-core"]
mcp = ["memex-mcp"]

[tool.uv.sources]
memex-common = { workspace = true }
memex-core = { workspace = true }
memex-mcp = { workspace = true }

[project.scripts]
memex = "memex_cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
